requirements.txt
python-dateutil
matplotlib
requests
edgar
numpy
alpaca_data_loader
pickle
fastapi>=0.68.0
uvicorn>=0.15.0
pandas>=1.3.0
plotly>=5.3.0
python-multipart>=0.0.5
